+++ src/command/mod.rs
@@ -2,12 +2,13 @@ use std::{process, thread, time::Duration};
 
 mod docker;
 
-pub fn run() {
+pub fn run(group: String) {
     info!(target: "lazymc-docker-proxy::command", "Starting server...");
     // Set a handler for SIGTERM
+    let cloned_group = group.clone();
     ctrlc::set_handler(move || {
         info!(target: "lazymc-docker-proxy::command", "Received SIGTERM, stopping server...");
-        docker::stop();
+        docker::stop(cloned_group.clone());
         process::exit(0);
     })
     .unwrap_or_else(|err| {
@@ -16,7 +17,7 @@ pub fn run() {
     });
 
     // Start the command
-    docker::start();
+    docker::start(group.clone());
 
     // Wait for SIGTERM
     loop {