+++ src/entrypoint/mod.rs
@@ -1,38 +1,21 @@
 mod config;
+mod docker;
 
-use std::{
-    env::var,
-    process::{self, exit, Command, ExitStatus},
-};
-use version_compare::Version;
+use config::Config;
+use std::process::{self, exit, ExitStatus};
 
 pub fn run() {
-    // lazymc dropped support minecraft servers with version less than 1.20.3
-    let is_legacy: bool = match var("PUBLIC_VERSION") {
-        Ok(version) => Version::from(version.as_ref()) < Version::from("1.20.3"),
-        Err(_) => false,
-    };
+    // Generate the lazymc config either from labels if set or from env
+    let config: Config;
+    let container_labels = docker::get_container_labels();
+    if container_labels.is_none() {
+        config = Config::from_env()
+    } else {
+        config = Config::from_container_labels(container_labels.unwrap());
+    }
 
-    // Generate the lazymc config file
-    config::generate(is_legacy);
-
-    // Create a new Command to run "lazymc start"
     info!(target: "lazymc-docker-proxy::entrypoint", "Starting lazymc process...");
-
-    // if is_legacy is true, run lazymc-legacy instead of lazymc
-    let program: &str = match is_legacy {
-        true => {
-            debug!(target: "lazymc-docker-proxy::entrypoint", "Running legacy version of lazymc as server protocol version is less than 1.20.3");
-            "lazymc-legacy"
-        },
-        false => {
-            debug!(target: "lazymc-docker-proxy::entrypoint", "Running newer version of lazymc as server protocol version is 1.20.3 or higher");
-            "lazymc"
-        },
-    };
-
-    let mut child: process::Child = Command::new(program)
-        .arg("start")
+    let mut child: process::Child = config.start_command()
         .spawn()
         .unwrap_or_else(|err| {
             error!(target: "lazymc-docker-proxy::entrypoint", "Failed to start lazymc process: {}", err);